allow embeddings index to carry arbitary models
