+++ api-examples/pretrain-paired.py
@@ -8,7 +8,7 @@ from torch.utils.data import DataLoader, TensorDataset, Dataset
 from eight_mile.utils import str2bool, write_json, Offsets, listify
 import glob
 from baseline.pytorch.embeddings import *
-import eight_mile.embeddings
+import baseline.embeddings
 from eight_mile.optz import *
 from eight_mile.pytorch.optz import *
 from eight_mile.pytorch.layers import TransformerEncoderStack, DenseStack, MultiHeadedAttention
@@ -511,7 +511,7 @@ def load_embed_and_vocab(token_type, reader, dataset, dataset_key, d_model, cach
         embeddings = {}
 
         # If we are not using chars, then use 'x' for both input and output
-        x_embedding = eight_mile.embeddings.load_embeddings('x',
+        x_embedding = baseline.embeddings.load_embeddings('x',
                                                             dsz=d_model,
                                                             known_vocab=vocab,
                                                             embed_type='positional')
