+++ baseline/pytorch/embeddings.py
@@ -1,4 +1,4 @@
-from eight_mile.embeddings import register_embeddings
+from baseline.embeddings import register_embeddings
 from eight_mile.pytorch.embeddings import *
