+++ mead/config/trec-bert-finetune-pytorch.json
@@ -1,0 +1,46 @@
+{
+  "task": "classify",
+  "basedir": "./trec-bert",
+  "batchsz": 10,
+  "modules": [
+    "embed_bert_pytorch"
+  ],
+  "features": [
+    {
+      "name": "bert",
+      "vectorizer": {
+        "type": "wordpiece1d",
+        "embed_file": "bert-base-uncased"
+      },
+      "embeddings": {
+        "type": "bert-pooled",
+        "embed_file": "bert-base-uncased",
+        "dsz": 768,
+        "trainable": true
+      }
+    }
+  ],
+  "preproc": {
+    "mxlen": 100
+  },
+  "backend": "pytorch",
+  "dataset": "trec",
+  "loader": {
+    "reader_type": "default"
+  },
+  "unif": 0.25,
+  "model": {
+    "model_type": "fine-tune"
+  },
+  "train": {
+    "epochs": 5,
+    "optim": "adamw",
+    "eta":  0.00001,
+    "weight_decay": 1.0e-8,
+    "early_stopping_metric": "acc",
+    "verbose": {
+      "console": true,
+      "file": "trec-cm.csv"
+    }
+  }
+}
